feat(audio): Enforce integer volume (0-100) and switch to quadratic taper for better audibility

+++ engine/config.py
@@ -71,7 +71,7 @@ class HotkeysConfig(BaseModel):
 class SoundsConfig(BaseModel):
     model_config = ConfigDict(extra="forbid")
     enabled: bool = True
-    volume: float = Field(default=30.0, ge=0.0, le=100.0)
+    volume: int = Field(default=30, ge=0, le=100)  # Percent (0-100)
     start_sound_path: str = r"C:\Windows\Media\Speech On.wav"
     stop_sound_path: str = r"C:\Windows\Media\Speech Off.wav"
